requirements.txt
bitshares==0.1.7
bitshares==0.1.7
Django==1.11.4
Fabric3==1.13.1.post1
coinbase==2.0.6
Fabric==1.13.2
djangorestframework==3.6.3
django-filter==1.0.4
psycopg2==2.7.1
django-cors-headers==2.1.0
django-extensions==1.8.1
